refactor: authenticate user
